% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mici.cuminc.R
\name{mici.cuminc}
\alias{mici.cuminc}
\title{Function to estimate the cumulative incidence from multiply imputed data}
\usage{
mici.cuminc(
  imp_obj = NULL,
  event_interest = 1,
  times = NULL,
  int.type = "Wilson",
  conf.width = 0.95,
  bayes_alpha = 0.8,
  bayes_beta = 1.2,
  bayes_samps = 1000
)
}
\arguments{
\item{imp_obj}{A list containing M imputed datasets free of censoring with columns
ftime and ftype containing the imputed event times and types, respectively. This 
is outputted from mici.impute().}

\item{event_interest}{The event indicator of the event of interest. The default is
1.}

\item{times}{A length l vector of times at which cumulative incidence estimates are 
requested. If none are provided, the cumulative incidence at every unique event time 
in the data will be returned.}

\item{int.type}{The type of uncertainty interval to use. Options include "Wald",
"Wilson", "Bayes", or "Wald_untrans". The default (and our recommendation for most 
settings) is "Wilson". Briefly, "Wald" is the log-log transformed Wald interval 
around the cumulative incidence. "Wilson" is a Wilson score interval around the 
cumulative incidence. "Bayes" is a Bayesian beta-binomial interval around the 
cumulative incidence. "Wald_untrans" is the untransformed Wald interval around the 
cumulative incidence (note that this is not constrained to fall in [0,1]).}

\item{conf.width}{The function will return a conf.width*100% uncertainty interval. 
The default is 0.95.}

\item{bayes_alpha}{If using the "Bayes" option for int.type, this is the alpha
shape parameter for the Beta prior on the cumulative incidence at all timepoints. 
The default is 0.8.}

\item{bayes_beta}{If using the "Bayes" option for int.type, this is the beta
shape parameter for the Beta prior on the cumulative incidence at all timepoints. 
The default is 1.2.}

\item{bayes_samps}{If using the "Bayes" option for int.type, this is the number 
of psoterior samples to draw for the beta-binomial interval on each imputation.
The default is 1000.}
}
\value{
The function returns a list with elements:
\describe{
 \item{n}{The sample size.}
 \item{time}{The length l vector of times.}
 \item{cuminc}{The cumulative incidence estimate at each timepoint.}
 \item{int.type}{The type of uncertainty interval being returned.}
 \item{conf.width}{The uncertainty interval width being returned.}
 \item{lower}{The lower bound of the uncertainty interval estimate at each timepoint.}
 \item{upper}{The upper bound of the uncertainty interval estimate at each timepoint.}
}
}
\description{
This function estimates the cumulative incidence from multiply imputed data
following the approach described in Chase et al. (2024+). Specifically, the 
function takes as input a list of M datasets in which all censoring has been
replaced with imputed event times and types. Then estimating the cumulative 
incidence at time t is just counting the number of events from the cause-of-interest
by time t and dividing by the sample size n.
}
\examples{
ftime <- sample(c(1:10), size = 100, replace = TRUE)
ftype <- sample(c(0:2), size = 100, replace = TRUE)

myimps <- mici.impute(ftime, ftype, M = 10, scheme = "KMI")
mycuminc <- mici.cuminc(myimps, int.type = "Wilson")

mydata <- data.frame("time" = sample(c(1:10), size = 100, replace = TRUE),
"type" = sample(c(0:2), size = 100, replace = TRUE))

myimps <- mici.impute(time, type, data = mydata, scheme = "RSI")
mycuminc <- mici.cuminc(myimps, int.type = "Wald")

}
